+++ app/newsPage/newsPage.component.ts
@@ -2,6 +2,7 @@ import {Component, OnInit} from '@angular/core';
 
 import {Article} from '../shared/article';
 import { ArticleService } from '../shared/article.service';
+import { UserService } from '../shared/user.service';
 import { ActivatedRoute } from '@angular/router';
 import { Router } from '@angular/router';
 
@@ -12,25 +13,39 @@ import { Router } from '@angular/router';
 })
 
 export class NewsPageComponent implements OnInit{
-    private articles;
     private currentArticles;
     private id: number;
 
-    constructor(private route: ActivatedRoute,private router: Router, private articleService: ArticleService){
-        this.articles = [];
+    model:boolean;
+
+    constructor(private route: ActivatedRoute,private router: Router, private articleService: ArticleService, private userService: UserService){
+        // this.articles = [];
     };
 
     ngOnInit(){
-        this.articles=this.articleService.articles;
         this.route.params.subscribe(params => {
             this.id = +params['id'];
         });
-        this.currentArticles=this.articleService.findArticle(this.id);
+        if(this.id==0){
+            this.currentArticles=this.articleService.getTemporaryArticle();
+        } else{
+            this.currentArticles=this.articleService.findArticle(this.id);
+        }
+        this.model=this.userService.displayEditButtons(this.currentArticles);
     }
     
     deleteArticle(){
         this.router.navigate(['/mainPage']);
-        this.articleService.deleteArticle(this.currentArticles);
+        if(this.id!==0){
+            this.articleService.deleteArticle(this.currentArticles);
+        } {
+            this.articleService.clearTemporaryArticle();
+            this.router.navigate(['/editorPage']);
+        }
+    }
+
+    directToEdit(){
+        this.router.navigate(['/editorPage', this.currentArticles.id]);
     }
     
 }
